🎉 Boom! We expanded functionality and added an API gateway, with more lambda functions for each.
