+++ app/add_feed/requirements.txt
@@ -1,0 +1,1 @@
+boto3
