+++ app/add_keyword/requirements.txt
@@ -1,0 +1,1 @@
+boto3
